Included known number as constraints to ilp problem

+++ solver/strategies/__init__.py
@@ -9,7 +9,7 @@ class AbstractStrategy:
     __metaclass__ = ABCMeta
 
     def __init__(self, field):
-        self.field = np.array(field)
+        self.field = np.array(field, dtype='int32')
         if field.shape != (9, 9):
             raise ValueError("Sudoku has field needs to be of shape (9,9)!")
         self.needed_numbers = set(range(1, 10))
